Backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
distro==1.9.0
dnspython==2.8.0
email-validator==2.0.0
fastapi==0.119.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
geojson==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
numpy==2.3.4
openai==2.7.1
overpass==0.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
preshed==3.0.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
smart_open==7.4.1
sniffio==1.3.1
# spacy==3.8.7
# spacy-legacy==3.0.12
# spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlmodel==0.0.27
srsly==2.5.1
starlette==0.48.0
thinc==8.3.6
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==2.0.0
minio==7.1.9
pypdf==3.17.0
python-docx==0.8.11
# https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0.tar.gz
# https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz